Initial RPC implementation
I gave up on the reflection idea, since that turned out to be a
lot harder than it sounds. instead, the "RPCHttpServer" class
will have hard-coded logic for each action we want to be able to
do.

The next thing to do is to get the applescript to defer to the
Python script when RD is running, and the Python script will
zippy zappy RPC the level over